# docs/openapi/repository-api-v1.json
{
  "components": {
    "schemas": {
...
      "ArtifactRefResponse": {
        "description": "Backend-agnostic artifact reference returned by the new fetch APIs.",
        "properties": {
          "checksum_algorithm": {
            "description": "Checksum algorithm for the immutable artifact.",
            "title": "Checksum Algorithm",
            "type": "string"
          },
          "checksum_digest": {
            "description": "Checksum digest for the immutable artifact.",
            "title": "Checksum Digest",
            "type": "string"
          },
          "download_path": {
            "description": "API path clients should call to stream the artifact.",
            "title": "Download Path",
            "type": "string"
          },
          "size_bytes": {
            "description": "Artifact size in bytes.",
            "title": "Size Bytes",
            "type": "integer"
          }
        },
        "required": [
          "checksum_algorithm",
          "checksum_digest",
          "size_bytes",
          "download_path"
        ],
        "title": "ArtifactRefResponse",
        "type": "object"
      },
...
      "ChecksumResponse": {
        "description": "Checksum metadata attached to a stored artifact.",
        "properties": {
          "algorithm": {
            "description": "Checksum algorithm used for the artifact.",
            "title": "Algorithm",
            "type": "string"
          },
          "digest": {
            "description": "Hex digest of the stored artifact.",
            "title": "Digest",
            "type": "string"
          }
        },
        "required": [
          "algorithm",
          "digest"
        ],
        "title": "ChecksumResponse",
        "type": "object"
      },
...
      "ErrorBody": {
        "description": "Error detail object for the public API error envelope.",
        "properties": {
          "code": {
            "description": "Stable machine-readable error code.",
            "title": "Code",
            "type": "string"
          },
          "details": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional structured metadata to help clients debug the error.",
            "title": "Details"
          },
          "message": {
            "description": "Human-readable summary of the failure.",
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
      "ErrorEnvelope": {
        "description": "Standardized error envelope shared across HTTP endpoints.",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody",
            "description": "Normalized error payload."
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
      "ExactSkillCoordinateRequest": {
        "additionalProperties": false,
        "description": "Exact immutable skill-version coordinate used by fetch and resolution reads.",
        "properties": {
          "skill_id": {
            "description": "Stable identifier of the requested skill.",
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9](?:[A-Za-z0-9._-]{0,127})$",
            "title": "Skill Id",
            "type": "string"
          },
          "version": {
            "description": "Exact immutable semantic version of the requested skill.",
            "pattern": "^(0|[1-9]\\d*)\\.(0|[1-9]\\d*)\\.(0|[1-9]\\d*)(?:-([0-9A-Za-z-]+(?:\\.[0-9A-Za-z-]+)*))?(?:\\+([0-9A-Za-z-]+(?:\\.[0-9A-Za-z-]+)*))?$",
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "skill_id",
          "version"
        ],
        "title": "ExactSkillCoordinateRequest",
        "type": "object"
      },
      "ExactSkillVersionResponse": {
        "description": "Exact immutable version metadata returned by the new fetch APIs.",
        "properties": {
          "artifact_ref": {
            "$ref": "#/components/schemas/ArtifactRefResponse",
            "description": "Backend-agnostic reference used to stream the immutable artifact."
          },
          "checksum": {
            "$ref": "#/components/schemas/ChecksumResponse",
            "description": "Checksum metadata for integrity checks."
          },
          "manifest": {
            "$ref": "#/components/schemas/SkillManifest",
            "description": "Validated manifest as stored by the service."
          },
          "published_at": {
            "description": "UTC timestamp when the version was published.",
            "format": "date-time",
            "title": "Published At",
            "type": "string"
          },
          "skill_id": {
            "description": "Stable identifier of the skill.",
            "title": "Skill Id",
            "type": "string"
          },
          "version": {
            "description": "Immutable semantic version of the skill.",
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "skill_id",
          "version",
          "manifest",
          "checksum",
          "artifact_ref",
          "published_at"
        ],
        "title": "ExactSkillVersionResponse",
        "type": "object"
      },
...
      "SkillFetchBatchItemResponse": {
        "description": "One ordered exact metadata batch result.",
        "properties": {
          "coordinate": {
            "$ref": "#/components/schemas/ExactSkillCoordinateRequest",
            "description": "Exact immutable coordinate as requested by the client."
          },
          "status": {
            "description": "Whether the requested coordinate was found.",
            "enum": [
              "found",
              "not_found"
            ],
            "title": "Status",
            "type": "string"
          },
          "version": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ExactSkillVersionResponse"
              },
              {
                "type": "null"
              }
            ],
            "description": "Returned immutable version metadata when the coordinate exists."
          }
        },
        "required": [
          "status",
          "coordinate"
        ],
        "title": "SkillFetchBatchItemResponse",
        "type": "object"
      },
      "SkillFetchBatchRequest": {
        "additionalProperties": false,
        "description": "Ordered batch request for exact immutable version metadata reads.",
        "properties": {
          "coordinates": {
            "description": "Ordered immutable coordinates to read exactly.",
            "items": {
              "$ref": "#/components/schemas/ExactSkillCoordinateRequest"
            },
            "maxItems": 100,
            "minItems": 1,
            "title": "Coordinates",
            "type": "array"
          }
        },
        "required": [
          "coordinates"
        ],
        "title": "SkillFetchBatchRequest",
        "type": "object"
      },
      "SkillFetchBatchResponse": {
        "description": "Ordered batch response for exact immutable metadata reads.",
        "properties": {
          "results": {
            "description": "Ordered per-coordinate fetch results preserving request order.",
            "items": {
              "$ref": "#/components/schemas/SkillFetchBatchItemResponse"
            },
            "title": "Results",
            "type": "array"
          }
        },
        "required": [
          "results"
        ],
        "title": "SkillFetchBatchResponse",
        "type": "object"
      },
      "SkillManifest": {
        "additionalProperties": false,
        "description": "Validated immutable skill manifest contract.",
        "properties": {
          "conflicts_with": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/RelationshipRef"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Immutable skill versions known to conflict with this version.",
            "title": "Conflicts With"
          },
          "depends_on": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/DependencyDeclaration"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Direct dependency declarations authored for this immutable version.",
            "title": "Depends On"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional human-readable summary of the skill.",
            "title": "Description"
          },
          "extends": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/RelationshipRef"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Other immutable skill versions this version extends.",
            "title": "Extends"
          },
          "name": {
            "description": "Human-readable skill name.",
            "maxLength": 200,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "overlaps_with": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/RelationshipRef"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Immutable skill versions with overlapping behavior or scope.",
            "title": "Overlaps With"
          },
          "schema_version": {
            "default": "1.0",
            "description": "Manifest schema version understood by this API.",
            "maxLength": 20,
            "minLength": 1,
            "title": "Schema Version",
            "type": "string"
          },
          "skill_id": {
            "description": "Stable catalog identifier for the skill.",
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9](?:[A-Za-z0-9._-]{0,127})$",
            "title": "Skill Id",
            "type": "string"
          },
          "tags": {
            "description": "Free-form tags used for categorization and discovery.",
            "items": {
              "type": "string"
            },
            "title": "Tags",
            "type": "array"
          },
          "version": {
            "description": "Immutable semantic version being published or fetched.",
            "pattern": "^(0|[1-9]\\d*)\\.(0|[1-9]\\d*)\\.(0|[1-9]\\d*)(?:-([0-9A-Za-z-]+(?:\\.[0-9A-Za-z-]+)*))?(?:\\+([0-9A-Za-z-]+(?:\\.[0-9A-Za-z-]+)*))?$",
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "skill_id",
          "version",
          "name"
        ],
        "title": "SkillManifest",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Registry-first API for immutable skill publication, indexed search candidate\nretrieval, direct relationship reads, exact version retrieval, artifact streaming,\nand version listing.\n\nThe server owns data-local registry operations only. Prompt interpretation,\nreranking, dependency solving, final selection, lock generation, and execution\nplanning remain client-owned behavior outside this API boundary.",
    "title": "aptitude-server Service",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/fetch/skill-versions:batch": {
      "post": {
        "description": "Return ordered exact immutable version metadata results for up to 100 requested coordinates. Each result is independent and preserves the original request order.",
        "operationId": "batchGetExactSkillVersionMetadata",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SkillFetchBatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "results": [
                    {
                      "coordinate": {
                        "skill_id": "python.lint",
                        "version": "1.2.3"
                      },
                      "status": "found",
                      "version": {
                        "artifact_ref": {
                          "checksum_algorithm": "sha256",
                          "checksum_digest": "c3ab8ff13720e8ad9047dd39466b3c8974e592c2fa383d4a3960714caef0c4f2",
                          "download_path": "/fetch/skills/python.lint/1.2.3/artifact",
                          "size_bytes": 123
                        },
                        "checksum": {
                          "algorithm": "sha256",
                          "digest": "c3ab8ff13720e8ad9047dd39466b3c8974e592c2fa383d4a3960714caef0c4f2"
                        },
                        "manifest": {
                          "conflicts_with": [],
                          "depends_on": [
                            {
                              "markers": [
                                "linux",
                                "gpu"
                              ],
                              "optional": true,
                              "skill_id": "python.base",
                              "version_constraint": ">=1.0.0,<2.0.0"
                            }
                          ],
                          "description": "Linting skill",
                          "extends": [
                            {
                              "skill_id": "python.base",
                              "version": "1.0.0"
                            }
                          ],
                          "name": "Python Lint",
                          "overlaps_with": [],
                          "schema_version": "1.0",
                          "skill_id": "python.lint",
                          "tags": [
                            "python",
                            "lint"
                          ],
                          "version": "1.2.3"
                        },
                        "published_at": "2026-03-10T08:30:00Z",
                        "skill_id": "python.lint",
                        "version": "1.2.3"
                      }
                    },
                    {
                      "coordinate": {
                        "skill_id": "python.missing",
                        "version": "9.9.9"
                      },
                      "status": "not_found"
                    }
                  ]
                },
                "schema": {
                  "$ref": "#/components/schemas/SkillFetchBatchResponse"
                }
              }
            },
            "description": "Ordered exact immutable metadata results returned successfully."
          },
          "422": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "INVALID_REQUEST",
                    "details": {
                      "errors": [
                        {
                          "input": "latest",
                          "loc": [
                            "path",
                            "version"
                          ],
                          "msg": "String should match pattern",
                          "type": "string_pattern_mismatch"
                        }
                      ]
                    },
                    "message": "Request validation failed."
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The request body or path parameters are invalid."
          }
        },
        "summary": "Batch fetch exact immutable version metadata",
        "tags": [
          "fetch"
        ]
      }
    },
    "/fetch/skills/{skill_id}/{version}": {
      "get": {
        "description": "Return the stored manifest, checksum, and a backend-agnostic artifact reference for one exact immutable skill version. This route does not inline artifact bytes.",
        "operationId": "getExactSkillVersionMetadata",
        "parameters": [
          {
            "description": "Stable identifier of the skill to fetch.",
            "in": "path",
            "name": "skill_id",
            "required": true,
            "schema": {
              "description": "Stable identifier of the skill to fetch.",
              "pattern": "^[A-Za-z0-9](?:[A-Za-z0-9._-]{0,127})$",
              "title": "Skill Id",
              "type": "string"
            }
          },
          {
            "description": "Exact immutable semantic version to fetch.",
            "in": "path",
            "name": "version",
            "required": true,
            "schema": {
              "description": "Exact immutable semantic version to fetch.",
              "pattern": "^(0|[1-9]\\d*)\\.(0|[1-9]\\d*)\\.(0|[1-9]\\d*)(?:-([0-9A-Za-z-]+(?:\\.[0-9A-Za-z-]+)*))?(?:\\+([0-9A-Za-z-]+(?:\\.[0-9A-Za-z-]+)*))?$",
              "title": "Version",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "artifact_ref": {
                    "checksum_algorithm": "sha256",
                    "checksum_digest": "c3ab8ff13720e8ad9047dd39466b3c8974e592c2fa383d4a3960714caef0c4f2",
                    "download_path": "/fetch/skills/python.lint/1.2.3/artifact",
                    "size_bytes": 123
                  },
                  "checksum": {
                    "algorithm": "sha256",
                    "digest": "c3ab8ff13720e8ad9047dd39466b3c8974e592c2fa383d4a3960714caef0c4f2"
                  },
                  "manifest": {
                    "conflicts_with": [],
                    "depends_on": [
                      {
                        "markers": [
                          "linux",
                          "gpu"
                        ],
                        "optional": true,
                        "skill_id": "python.base",
                        "version_constraint": ">=1.0.0,<2.0.0"
                      }
                    ],
                    "description": "Linting skill",
                    "extends": [
                      {
                        "skill_id": "python.base",
                        "version": "1.0.0"
                      }
                    ],
                    "name": "Python Lint",
                    "overlaps_with": [],
                    "schema_version": "1.0",
                    "skill_id": "python.lint",
                    "tags": [
                      "python",
                      "lint"
                    ],
                    "version": "1.2.3"
                  },
                  "published_at": "2026-03-10T08:30:00Z",
                  "skill_id": "python.lint",
                  "version": "1.2.3"
                },
                "schema": {
                  "$ref": "#/components/schemas/ExactSkillVersionResponse"
                }
              }
            },
            "description": "Exact immutable skill version metadata returned successfully."
          },
          "404": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "SKILL_VERSION_NOT_FOUND",
                    "details": {
                      "skill_id": "python.lint",
                      "version": "9.9.9"
                    },
                    "message": "Skill version not found: python.lint@9.9.9"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The requested immutable `skill_id@version` does not exist."
          },
          "422": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "INVALID_REQUEST",
                    "details": {
                      "errors": [
                        {
                          "input": "latest",
                          "loc": [
                            "path",
                            "version"
                          ],
                          "msg": "String should match pattern",
                          "type": "string_pattern_mismatch"
                        }
                      ]
                    },
                    "message": "Request validation failed."
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The request body or path parameters are invalid."
          }
        },
        "summary": "Fetch exact immutable version metadata",
        "tags": [
          "fetch"
        ]
      }
    },
    "/fetch/skills/{skill_id}/{version}/artifact": {
      "get": {
        "description": "Stream the immutable artifact bytes for one exact `skill_id@version`. The service verifies checksum integrity before returning the payload.",
        "operationId": "streamExactSkillArtifact",
        "parameters": [
          {
            "description": "Stable identifier of the skill to fetch.",
            "in": "path",
            "name": "skill_id",
            "required": true,
            "schema": {
              "description": "Stable identifier of the skill to fetch.",
              "pattern": "^[A-Za-z0-9](?:[A-Za-z0-9._-]{0,127})$",
              "title": "Skill Id",
              "type": "string"
            }
          },
          {
            "description": "Exact immutable semantic version to fetch.",
            "in": "path",
            "name": "version",
            "required": true,
            "schema": {
              "description": "Exact immutable semantic version to fetch.",
              "pattern": "^(0|[1-9]\\d*)\\.(0|[1-9]\\d*)\\.(0|[1-9]\\d*)(?:-([0-9A-Za-z-]+(?:\\.[0-9A-Za-z-]+)*))?(?:\\+([0-9A-Za-z-]+(?:\\.[0-9A-Za-z-]+)*))?$",
              "title": "Version",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              },
              "application/octet-stream": {}
            },
            "description": "Immutable artifact bytes streamed successfully."
          },
          "404": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "SKILL_VERSION_NOT_FOUND",
                    "details": {
                      "skill_id": "python.lint",
                      "version": "9.9.9"
                    },
                    "message": "Skill version not found: python.lint@9.9.9"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The requested immutable `skill_id@version` does not exist."
          },
          "422": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "INVALID_REQUEST",
                    "details": {
                      "errors": [
                        {
                          "input": "latest",
                          "loc": [
                            "path",
                            "version"
                          ],
                          "msg": "String should match pattern",
                          "type": "string_pattern_mismatch"
                        }
                      ]
                    },
                    "message": "Request validation failed."
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The request body or path parameters are invalid."
          },
          "500": {
            "content": {
              "application/json": {
                "examples": {
                  "artifact_storage_failure": {
                    "value": {
                      "error": {
                        "code": "ARTIFACT_STORAGE_FAILURE",
                        "message": "Artifact storage failed during publish."
                      }
                    }
                  },
                  "integrity_check_failed": {
                    "value": {
                      "error": {
                        "code": "INTEGRITY_CHECK_FAILED",
                        "details": {
                          "skill_id": "python.lint",
                          "version": "1.2.3"
                        },
                        "message": "Integrity check failed for skill version: python.lint@1.2.3"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Artifact retrieval or integrity verification failed."
          }
        },
        "summary": "Stream one immutable artifact",
        "tags": [
          "fetch"
        ]
      }
    },
...
  }
}